pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.127.0",
    "passlib[argon2,bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "python-ulid>=3.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
